Define 6 distinct visualizations pages:         - affinity plotting         - affinity chord plotting         - covariance network plotting         - community detection plotting         - nldr plotting
Each of these will orchestrate the a full visualization page, that is
selected via the initial drop-down menu.

Each of these visualizations will be comprised of multiple components,
drawing from the new components/ directory. Currently, existing shared
code that is not in one of the main visualization modules has been
added, but more will be moved into the components directory from the
*plotting.js files.

A separate utilities/ directory was made for abstract support functions
for math, html manipulation, etc.